+++ source/core/core.h
@@ -35,13 +35,13 @@ class Core {
         Core() : periodSize(2048), samplingRate(44100), generatorCounter(0), proc(nullptr), out(nullptr) {};
 
         int32_t setProcessor(processor::types procType, processor::ConfigData *cfgData = nullptr);
-        uint16_t addGenerator(generator::types genType, generator::ConfigData *cfgData = nullptr);
+        int16_t addGenerator(generator::types genType, generator::ConfigData *cfgData = nullptr);
         int32_t setOutput(output::types outType, output::ConfigData *cfgData = nullptr);
         uint32_t getPeriodSize() {return periodSize;}
         void setPeriodSize(uint32_t periodSize);
         uint32_t getSamplingRate() {return samplingRate;}
         void setSamplingRate(uint32_t samplingRate) {samplingRate = samplingRate;}
-        int32_t addSource(processor::SourceConfigData *srcData = nullptr);
+        int16_t addSource(processor::SourceConfigData *srcData = nullptr);
         void render();
         void generatorConfig(int gid, generator::ConfigData *configData);  //optional, in case one wishes to alter the generators default values
         void processorConfig(processor::ConfigData *configData);         //optional, in case one wishes to alter the generators default values
